requirements.txt
asgiref==3.4.1
asn1crypto==0.24.0
backcall==0.2.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.3.7
cryptography==2.1.4
decorator==5.1.1
Django==3.2.15
django-common==0.1.51
djangorestframework==3.13.1
httplib2==0.9.2
idna==2.6
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
keyring==10.6.0
keyrings.alt==3.0
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
macaroonbakery==1.1.3
Mako==1.0.7
netifaces==0.10.4
numpy==1.19.5
oauth==1.0.1
olefile==0.45.1
pandas==1.1.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==5.1.0
prompt-toolkit==3.0.31
protobuf==3.0.0
ptyprocess==0.7.0
pycrypto==2.6.1
Pygments==2.13.0
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-dateutil==2.8.2
python-debian==0.1.32
pytz==2022.2.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
SecretStorage==2.3.1
simplejson==3.13.2
six==1.16.0
sqlparse==0.4.2
traitlets==4.3.3
typing-extensions==4.1.1
urllib3==1.22
wadllib==1.3.2
wcwidth==0.2.5